% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{AnnProc}
\alias{AnnProc}
\title{Function to process one single annotation}
\usage{
AnnProc(antn, nRecords = 3, segments = 1:8, raw = FALSE,
  parBarStats = FALSE)
}
\arguments{
\item{antn}{One Annotation Records}

\item{nRecords}{Num of Records per annotation}

\item{segments}{A vector indicating the PAR bar segments selected}

\item{raw}{Raw outputdata}

\item{parBarStats}{include some stastistics of the PAR bar in the output}
}
\value{
a dataframe with the means of each record or raw data if selected, if the annotation is not 
consistent, returns null
}
\description{
Function to process one single annotation
}
