% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{CeptProc}
\alias{CeptProc}
\title{Process the ceptometer file as a dataframe}
\usage{
CeptProc(ceptData, segments = 1:8, raw = TRUE,
  parBarStatistics = FALSE, recordOrder = NA, trialName = NA)
}
\arguments{
\item{ceptData}{A dataframe readed from the ceptometer output file}

\item{segments}{A vector indicating the PAR bar segments to be extracted}

\item{raw}{A boolean to output data in raw format}

\item{parBarStatistics}{A boolean, include some stastistics of the PARbar in the output}

\item{recordOrder}{A string of the expected record order, i.e 'ABA' means 'ABV','BLW','AVB'}

\item{trialName}{A string in regex format indicating the name of the annotation to be extracted}
}
\value{
A data frame containing all or a subset of the annotations
}
\description{
Process the ceptometer file as a dataframe
}
