% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{getPARBarStats}
\alias{getPARBarStats}
\title{Calculate some stats of the PAR from the raw anotation}
\usage{
getPARBarStats(parMat, strat)
}
\arguments{
\item{parMat}{a Matrix of the PAR bar segments}

\item{strat}{a vector with the name of each strata, i.e ARRIBA,REFLEC, etc.}
}
\value{
A list with the calculated params
}
\description{
Calculate some stats of the PAR from the raw anotation
}
