% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{getPARBarStatsRaw}
\alias{getPARBarStatsRaw}
\title{Calculate PAR bar statistics for each record's raw data}
\usage{
getPARBarStatsRaw(rawRecords, rejectOutliers = FALSE)
}
\arguments{
\item{rawRecords}{a dataframe returned by AnnProc setting the raw flag as TRUE}

\item{rejectOutliers}{a boolean flag to reject outliers in each record and calcalate the stats}
}
\value{
a dataframe with the raw data and the calculated statistics
}
\description{
Calculate PAR bar statistics for each record's raw data
}
