% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{recordsOutlierReject}
\alias{recordsOutlierReject}
\title{Reject outliers in each record}
\usage{
recordsOutlierReject(records)
}
\arguments{
\item{records}{a dataframe returned by AnnProc setting the raw flag as TRUE}
}
\value{
a dataframe without outliers in each record
}
\description{
Reject outliers in each record
}
