% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{SubsetAnn}
\alias{SubsetAnn}
\title{Subsetting one or more annotations from the ceptometer file given as a dataframe}
\usage{
SubsetAnn(df, tName, nRecords = 3, segments = 1:8, asDf = TRUE,
  raw = FALSE, parBarStats = FALSE)
}
\arguments{
\item{df}{A dataframe readed from the ceptometer output file}

\item{tName}{A string in regex format indicating the name of the annotation(s) to be extracted}

\item{nRecords}{Num of Records per annotation}

\item{segments}{A vector indicating the PAR bar segments selected}

\item{asDf}{function output as dataframe or list of dataframes}

\item{raw}{Raw outputdata}

\item{parBarStats}{include some stastistics of the PAR bar in the output}
}
\value{
A data frame containing a subset of the annotations
}
\description{
Subsetting one or more annotations from the ceptometer file given as a dataframe
}
