% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cept_proc_funtions.R
\name{SubsetByBnd}
\alias{SubsetByBnd}
\title{Extract one annotation from dataframe given the annotation boundaries
This fucntion calls directly the AnnProc function}
\usage{
SubsetByBnd(bnd, df, nRecords, segments, raw, parBarStats)
}
\arguments{
\item{df}{ceptometer file dataframe}

\item{nRecords}{Num of Records per annotation}

\item{segments}{A vector indicating the PAR bar segments selected}

\item{raw}{Raw outputdata}

\item{parBarStats}{include some stastistics of the PAR bar in the output}

\item{bdn}{A dataframe containing the name annotation names and their boundaries}

\item{antn}{One Annotation Records}
}
\value{
a dataframe defined by AnnProc function
}
\description{
Extract one annotation from dataframe given the annotation boundaries
This fucntion calls directly the AnnProc function
}
